requirements.txt
pytorch-lightning
lightning-bolts
wandb
torchvision